+++ tests/vvrest_test_suite.py
@@ -2,6 +2,10 @@ import unittest
 from .setup import setup_test_suite
 from .utilities import get_vault_object
 from .auth_service_test import AuthServiceTest
+from .config_service_test import ConfigServiceTest
+from .docapi_document_service_test import DocApiDocumentServiceTest
+from .docapi_file_service_test import DocApiFileServiceTest
+from .docapi_search_service_test import DocApiSearchServiceTest
 from .document_service_test import DocumentServiceTest
 from .email_service_test import EmailServiceTest
 from .file_service_test import FileServiceTest
@@ -16,6 +20,10 @@ setup_test_suite()  # setup VVRestTestSuite
 
 vault = get_vault_object()
 AuthServiceTest.vault = vault
+ConfigServiceTest.vault = vault
+DocApiDocumentServiceTest.vault = vault
+DocApiFileServiceTest.vault = vault
+DocApiSearchServiceTest.vault = vault
 DocumentServiceTest.vault = vault
 EmailServiceTest.vault = vault
 FileServiceTest.vault = vault